Added JavaDocs
