--- src/main/java/Duke/commands/IncorrectCommand.java
@@ -1,18 +1,0 @@
-package Duke.commands;
-
-import Duke.storage.Storage;
-import Duke.tasks.TaskList;
-import Duke.ui.UI;
-
-public class IncorrectCommand extends Command {
-
-    @Override
-    public void execute(TaskList tasks, UI ui, Storage storage) {
-        ui.incorrectCommandMessage();
-    }
-
-    @Override
-    public boolean isExit() {
-        return false;
-    }
-}
