+++ src/main/java/duke/commands/AddDeadlineCommand.java
@@ -1,10 +1,10 @@
-package Duke.commands;
+package duke.commands;
 
-import Duke.exceptions.DukeException;
-import Duke.storage.Storage;
-import Duke.tasks.Deadline;
-import Duke.tasks.TaskList;
-import Duke.ui.UI;
+import duke.exceptions.DukeException;
+import duke.storage.Storage;
+import duke.tasks.Deadline;
+import duke.tasks.TaskList;
+import duke.ui.UI;
 
 import java.time.LocalDateTime;
 
@@ -19,6 +19,13 @@ public class AddDeadlineCommand extends Command {
         this.input = input;
     }
 
+    /**
+     * Adds a new Deadline task to the task list.
+     *
+     * @param taskList the destination task list for the deadline to be added
+     * @param ui the ui to display message after the task is added
+     * @param storage the storage to handle storing of the new task list
+     */
     @Override
     public void execute(TaskList taskList, UI ui, Storage storage) {
         String taskDesc = input.substring(9, input.indexOf('/') - 1);
@@ -33,10 +40,15 @@ public class AddDeadlineCommand extends Command {
         }
         Deadline deadline = new Deadline(taskDesc, date.toString());
         taskList.addTask(deadline);
-        ui.addTaskMessage(deadline, taskList.size());
+        ui.addTaskMessage(deadline, taskList.getSize());
         storage.store(taskList);
     }
 
+    /**
+     * Prevents the program from terminating in Duke.run().
+     *
+     * @return False as this is not the 'exit' command
+     */
     @Override
     public boolean isExit() {
         return false;
