+++ src/main/java/duke/storage/Storage.java
@@ -1,31 +1,38 @@
-package Duke.storage;
+package duke.storage;
 
-import Duke.tasks.*;
+
+import duke.tasks.Task;
+import duke.tasks.TaskList;
 
 import java.io.*;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
 
-/**
- * Represents the file used to store the TaskList
- */
 public class Storage {
 
     /** Default file path used if the user does not provide the file name. */
     static String path = Paths.get(".").toAbsolutePath().toString();
-    private static final String DEFAULT_FILEPATH = path.substring(0, path.length() - 1) + "src/main/";
+    private static final String RELATIVE_FILEPATH = path.substring(0, path.length() - 1) + "src/main/";
 
     private String filePath = "";
 
+    /**
+     * Saves file to a default path if one was not provided by the user.
+     */
     public Storage() {
-        this.filePath = DEFAULT_FILEPATH + "data/tasks.txt";
+        this.filePath = RELATIVE_FILEPATH + "data/tasks.txt";
     }
 
+    /**
+     * Saves the file to the file path specified by the user.
+     *
+     * @param filePath path for the file to be stored
+     */
     @SuppressWarnings("ResultOfMethodCallIgnored")
     public Storage(String filePath) {
         try {
-            this.filePath = DEFAULT_FILEPATH + filePath;
+            this.filePath = RELATIVE_FILEPATH + filePath;
             File file = new File(this.filePath);
             file.createNewFile();
         } catch (IOException e) {
@@ -33,6 +40,11 @@ public class Storage {
         }
     }
 
+    /**
+     * Stores the given task list to the given path.
+     *
+     * @param toStore the task list to be stored
+     */
     public void store(TaskList toStore) {
         try {
             BufferedWriter writer = new BufferedWriter(new FileWriter(this.filePath));
@@ -47,6 +59,11 @@ public class Storage {
     }
 
 
+    /**
+     * Loads the stored task list from the specified file path.
+     *
+     * @return an ArrayList<Task> to be passed as parameters for the TaskList constructor
+     */
     public ArrayList<Task> load() {
         ArrayList<Task> result = new ArrayList<>();
         try {
