+++ src/test/java/jUnitTests/GetTaskTypeTest.java
@@ -1,8 +1,8 @@
 package jUnitTests;
 
-import Duke.tasks.Deadline;
-import Duke.tasks.Event;
-import Duke.tasks.Todo;
+import duke.tasks.Deadline;
+import duke.tasks.Event;
+import duke.tasks.Todo;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
